requirements.txt
flask
flask-cors
requests
beautifulsoup4
PyPDF2
python-dotenv
markitdown
